color theme is loaded upon color palette
