+++ raynian-fe/src/components/ui/background-customizer/BgCustomizerMenu.jsx
@@ -162,7 +162,7 @@ const BgCustomizerMenu = () => {
             />
           </div>
 
-          <motion.div
+          {/* <motion.div
             className="mt-4 flex justify-center"
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
@@ -179,9 +179,9 @@ const BgCustomizerMenu = () => {
                 {showForm ? <FaAngleUp /> : <FaAngleDown />}
               </motion.span>
             </button>
-          </motion.div>
+          </motion.div> */}
 
-          <motion.div
+          {/* <motion.div
             initial={{ height: 0 }}
             animate={{ height: showForm ? "auto" : 0 }}
             transition={{ duration: 0.5 }}
@@ -287,7 +287,7 @@ const BgCustomizerMenu = () => {
                 </div>
               </div>
             </div>
-          </motion.div>
+                  </motion.div> */}
         </div>
       </div>
     </ColorPickerStatusProvider>
